figure labels and 6D/8D figures
trying to use the [0,2) range for consistency but it makes the
calculations more annoying when you have to specify every coordinate in
terms of 1/2, will have to think how I want to do that.

Main update is the img/8d.py, adds a view of a full 8D constellation
design using different shapes/colours/sizes, it is a little hard to
understand but not terribly.

I'd like to update the 8D to have more clear indicators from each bit
instead of pairing up the bits to get two 4-way choices but can't think
of a clear way to deal yet.